PUBDEV-6497: Introduced Writer as an extra level of abstraction.
